requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
openai>=1.12.0
werkzeug==3.0.1
python-dotenv==1.0.0
httpx>=0.27.0

